refactor: keep ProxyResponse and UnwrapProxyLinkToken untouched
Move all paced streaming logic into ProxyResponsePaced in a new
http_paced.go file. Revert http.go to upstream, revert cosmetic
changes to CreateProxyLink and UnwrapProxyLinkToken in store.go.
Only additive qBit-specific code remains in shared.
