+++ internal/endpoint/proxy.go
@@ -30,22 +30,22 @@ func handleProxyLinkAccess(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	info, err := shared.UnwrapProxyLinkToken(encodedToken)
+	user, link, headers, tunnelType, err := shared.UnwrapProxyLinkToken(encodedToken)
 	if err != nil {
 		SendError(w, r, err)
 		return
 	}
 
-	if info.Headers != nil {
-		for k, v := range info.Headers {
+	if headers != nil {
+		for k, v := range headers {
 			r.Header.Set(k, v)
 		}
 	}
 
-	if isGetReq && info.User != "" {
-		cpStore := contentProxyConnectionStore.WithScope(info.User)
+	if isGetReq && user != "" {
+		cpStore := contentProxyConnectionStore.WithScope(user)
 
-		if limit := config.ContentProxyConnectionLimit.Get(info.User); limit > 0 {
+		if limit := config.ContentProxyConnectionLimit.Get(user); limit > 0 {
 			activeConnectionCount, err := cpStore.Count()
 			if err != nil {
 				ctx.Log.Error("[proxy] failed to count connections", "error", err)
@@ -55,40 +55,37 @@ func handleProxyLinkAccess(w http.ResponseWriter, r *http.Request) {
 			}
 		}
 
-		if err := cpStore.Set(ctx.RequestId, contentProxyConnection{IP: core.GetRequestIP(r), Link: info.Link}); err != nil {
+		if err := cpStore.Set(ctx.RequestId, contentProxyConnection{IP: core.GetRequestIP(r), Link: link}); err != nil {
 			ctx.Log.Error("[proxy] failed to record connection", "error", err)
 		} else {
 			defer cpStore.Del(ctx.RequestId)
 		}
 	}
 
-	// For qBittorrent files, provide a progress callback so the proxy can
-	// pace streaming to match the download — preventing garbage bytes from
-	// pre-allocated regions while allowing seamless streaming-while-downloading.
+	// For qBittorrent files, use the paced proxy to match streaming to the
+	// download progress — preventing garbage bytes from pre-allocated regions.
 	// The torrent hash and file index are passed as URL query params by
 	// GenerateStremThruLink, keeping qBit-specific concerns out of the
 	// core proxy token.
-	var safeBytesFn shared.SafeBytesFunc
-	var isRangeAvailFn shared.IsRangeAvailableFunc
+	var bytesWritten int64
 	qbitHash := r.URL.Query().Get("qbit_hash")
-	qbitFileIdx := 0
-	if fidxStr := r.URL.Query().Get("qbit_fidx"); fidxStr != "" {
-		if v, err := strconv.Atoi(fidxStr); err == nil {
-			qbitFileIdx = v
-		}
-	}
 	if qbitHash != "" {
-		safeBytesFn = func() (int64, int64, bool) {
-			safe, fileSize, done, err := shared.GetQbitSafeBytes(info.User, qbitHash, qbitFileIdx)
+		qbitFileIdx := 0
+		if fidxStr := r.URL.Query().Get("qbit_fidx"); fidxStr != "" {
+			if v, err := strconv.Atoi(fidxStr); err == nil {
+				qbitFileIdx = v
+			}
+		}
+		safeBytesFn := func() (int64, int64, bool) {
+			safe, fileSize, done, err := shared.GetQbitSafeBytes(user, qbitHash, qbitFileIdx)
 			if err != nil {
-				// On error, assume fully downloaded to avoid blocking forever.
 				ctx.Log.Warn("[proxy] failed to get qbit safe bytes, assuming done", "error", err)
 				return 0, 0, true
 			}
 			return safe, fileSize, done
 		}
-		isRangeAvailFn = func(start, end int64) bool {
-			avail, err := shared.IsQbitFileRangeAvailable(info.User, qbitHash, qbitFileIdx, start, end)
+		isRangeAvailFn := func(start, end int64) bool {
+			avail, err := shared.IsQbitFileRangeAvailable(user, qbitHash, qbitFileIdx, start, end)
 			if err != nil {
 				ctx.Log.Warn("[proxy] failed to check range availability", "error", err)
 				return false
@@ -99,10 +96,11 @@ func handleProxyLinkAccess(w http.ResponseWriter, r *http.Request) {
 			return avail
 		}
 		ctx.Log.Debug("[proxy] streaming with qbit progress awareness", "hash", qbitHash, "fileIdx", qbitFileIdx)
+		bytesWritten, err = shared.ProxyResponsePaced(w, r, link, tunnelType, safeBytesFn, isRangeAvailFn)
+	} else {
+		bytesWritten, err = shared.ProxyResponse(w, r, link, tunnelType)
 	}
-
-	bytesWritten, err := shared.ProxyResponse(w, r, info.Link, info.TunnelType, safeBytesFn, isRangeAvailFn)
-	ctx.Log.Info("[proxy] connection closed", "user", info.User, "size", util.ToSize(bytesWritten), "error", err)
+	ctx.Log.Info("[proxy] connection closed", "user", user, "size", util.ToSize(bytesWritten), "error", err)
 }
 
 type proxifyLinksData struct {
