+++ internal/stremio/addon/client.go
@@ -283,7 +283,7 @@ func (c Client) ProxyResource(w http.ResponseWriter, r *http.Request, params *Pr
 	}
 	adjustClientIPHeader(params.Ctx, params.ClientIP, r)
 	w.Header().Del("Access-Control-Allow-Origin")
-	shared.ProxyResponse(w, r, params.BaseURL.JoinPath(path).String(), config.TUNNEL_TYPE_AUTO, nil, nil)
+	shared.ProxyResponse(w, r, params.BaseURL.JoinPath(path).String(), config.TUNNEL_TYPE_AUTO)
 }
 
 func NormalizeManifestURL(manifestUrl string) (string, error) {
